docs: render the search palette only once
@theme/SearchBar is mounted twice on docs pages: once from the navbar's
search item and once in the swizzled docs sidebar. Each instance
installed its own global Cmd/Ctrl+K listener and portaled its own
overlay, so the shortcut opened two stacked palettes and the backdrop
had to be clicked twice to fully dismiss them.

Move the open state and the keyboard shortcut into a shared,
module-level store so there is a single listener, reduce SearchBar to
just the trigger button, and render the overlay from one SearchPalette
mounted once via a swizzled Root. Both trigger buttons remain, but there
is now only ever one palette.
